refactor controller/httprest
