+++ internal/controller/httprest/routes.go
@@ -1,8 +1,9 @@
-package metrics
+package httprest
 
 import (
 	"net/http"
 
+	"github.com/crazylazyowl/metrics-tpl/internal/controller/httprest/api"
 	"github.com/crazylazyowl/metrics-tpl/internal/controller/httprest/middleware"
 	"github.com/crazylazyowl/metrics-tpl/internal/usecase/metrics"
 
@@ -10,16 +11,8 @@ import (
 )
 
 func NewRouter(metrics *metrics.Usecase) http.Handler {
-	api := NewAPI(metrics)
-
 	r := chi.NewRouter()
-
 	r.Use(middleware.Logger)
-
-	r.Get("/", api.GetMetrics)
-
-	r.Get("/value/{type}/{metric}", api.GetMetric)
-	r.Post("/update/{type}/{metric}/{value}", api.UpdateMetric)
-
+	r.Mount("/", api.NewMetricsRouter(metrics))
 	return r
 }
